Delete unneeded signup component file
